refactor: moved last functions to global utils
