+++ cognite_toolkit/_cdf_tk/data_classes/_migration_yaml.py
@@ -12,8 +12,7 @@ from rich.panel import Panel
 
 from cognite_toolkit._cdf_tk.constants import COGNITE_MODULES
 from cognite_toolkit._cdf_tk.exceptions import ToolkitMigrationError
-from cognite_toolkit._cdf_tk.templates._utils import iterate_modules
-from cognite_toolkit._cdf_tk.utils import load_yaml_inject_variables
+from cognite_toolkit._cdf_tk.utils import iterate_modules, load_yaml_inject_variables
 from cognite_toolkit._version import __version__ as current_version
