--- cognite_toolkit/_cdf_tk/templates/__init__.py
@@ -1,3 +1,0 @@
-from ._utils import iterate_modules, module_from_path, resource_folder_from_path
-
-__all__ = ["iterate_modules", "resource_folder_from_path", "module_from_path"]
