--- cognite_toolkit/_cdf_tk/templates/_utils.py
@@ -1,65 +1,0 @@
-"""Local utils for the templates module"""
-
-from __future__ import annotations
-
-from collections.abc import Iterator
-from pathlib import Path
-from typing import Literal, overload
-
-from cognite_toolkit._cdf_tk.constants import EXCL_FILES
-from cognite_toolkit._cdf_tk.loaders import LOADER_BY_FOLDER_NAME
-
-
-def iterate_modules(root_dir: Path) -> Iterator[tuple[Path, list[Path]]]:
-    """Iterate over all modules in the project and yield the module directory and all files in the module.
-
-    Args:
-        root_dir (Path): The root directory of the project
-
-    Yields:
-        Iterator[tuple[Path, list[Path]]]: A tuple containing the module directory and a list of all files in the module
-
-    """
-    if not root_dir.exists():
-        return
-    for module_dir in root_dir.iterdir():
-        if not module_dir.is_dir():
-            continue
-        sub_directories = [path for path in module_dir.iterdir() if path.is_dir()]
-        is_any_resource_directories = any(dir.name in LOADER_BY_FOLDER_NAME for dir in sub_directories)
-        if sub_directories and is_any_resource_directories:
-            # Module found
-            yield module_dir, [path for path in module_dir.rglob("*") if path.is_file() and path.name not in EXCL_FILES]
-            # Stop searching for modules in subdirectories
-            continue
-        yield from iterate_modules(module_dir)
-
-
-@overload
-def module_from_path(path: Path, return_resource_folder: Literal[True]) -> tuple[str, str]: ...
-
-
-@overload
-def module_from_path(path: Path, return_resource_folder: Literal[False] = False) -> str: ...
-
-
-def module_from_path(path: Path, return_resource_folder: bool = False) -> str | tuple[str, str]:
-    """Get the module name from a path"""
-    if len(path.parts) == 1:
-        raise ValueError("Path is not a module")
-    last_folder = path.parts[1]
-    for part in path.parts[1:]:
-        if part in LOADER_BY_FOLDER_NAME:
-            if return_resource_folder:
-                return last_folder, part
-            return last_folder
-        last_folder = part
-    raise ValueError("Path is not part of a module")
-
-
-def resource_folder_from_path(path: Path) -> str:
-    """Get the resource_folder from a path"""
-    for part in path.parts:
-        if part in LOADER_BY_FOLDER_NAME:
-            return part
-    raise ValueError("Path does not contain a resource folder")
