+++ tests/tests_unit/test_cdf_tk/test_templates.py
@@ -16,11 +16,12 @@ from cognite_toolkit._cdf_tk.data_classes import (
     SystemYAML,
 )
 from cognite_toolkit._cdf_tk.loaders import LOADER_BY_FOLDER_NAME
-from cognite_toolkit._cdf_tk.templates import (
+from cognite_toolkit._cdf_tk.utils import (
+    YAMLComment,
+    flatten_dict,
     iterate_modules,
     module_from_path,
 )
-from cognite_toolkit._cdf_tk.utils import YAMLComment, flatten_dict
 from tests.tests_unit.data import PYTEST_PROJECT
 from tests.tests_unit.test_cdf_tk.constants import BUILD_DIR
